+++ src/main/java/com/thaichicken/cinemabooking/repository/MovieRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MovieRepository extends JpaRepository<MovieEntity, Long> {
+public interface MovieRepository extends JpaRepository<MovieEntity, Integer> {
     List<MovieEntity> findByDirector(String director);
 }
