+++ src/main/java/com/thaichicken/cinemabooking/repository/ShowTimeSeatRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ShowTimeSeatRepository extends JpaRepository<ShowTimeEntity, Long> {
+public interface ShowTimeSeatRepository extends JpaRepository<ShowTimeEntity, Integer> {
 }
